requirements-dev.txt
uvicorn==0.15.0
pydantic==1.8.2
fastapi==0.68.1
python-dotenv==0.19.0

# for linting, testing and coverage
pylint==2.11.1
pytest==6.2.5
pre-commit==2.15.0
coverage==5.5
lxml==4.6.3
mypy==0.910

# Workaround https://github.com/tiangolo/fastapi/issues/2595
requests==2.26.0
